* Fixed automatic updating of UI based on style under the cursor in RTF mode. * Fixed color selection in RTF mode.

+++ src/QuickPad.Standard.Mvvm/ViewModels/DocumentViewModel.cs
@@ -126,7 +126,7 @@ namespace QuickPad.Mvvm.ViewModels
                     Document.CurrentFontName =
                         isRtf ? Settings.DefaultRtfFont : Settings.DefaultFont;
                     Document.CurrentFontSize =
-                        isRtf ? Settings.DefaultFontRtfSize : Settings.DefaultFontSize;
+                        isRtf ? (float)Settings.DefaultFontRtfSize : (float)Settings.DefaultFontSize;
                     Document.CurrentWordWrap =
                         isRtf ? Settings.RtfWordWrap : Settings.WordWrap;
                     Document.GetOptions =
